Adding MultiMonitor Sample
Regularly Industrial Applications/SCADA Applications need the Viewer to be on multiple monitors. This sample shows how to create a multi-monitor environment based on CLIENT IP Address. Imagine that your clients have different monitor configurations (one has 3 monitors, the other has 4, there are others with only 1). A script on the server side, will validate which IP Address is trying to connect and resolve to open 1,2,3 or 4 monitors on the client.